+++ pages/about.js
@@ -6,11 +6,8 @@ import Container from '@/components/containers/Container';
 import { Row } from '@/components/containers/Row';
 import Member from '@/components/containers/Member';
 import { whoWeAre } from '@/utils/about';
-import { useTheme } from 'styled-components';
 
 export default function AboutUs() {
-  const theme = useTheme();
-
   return (
     <div className='about-us'>
       <RevealContentContainer>
@@ -31,9 +28,8 @@ export default function AboutUs() {
           rowOrder='row'
           image='/images/svg/square-brackets.svg'
           altTag=''
-          color={theme.colors.primaryContent}
-          bgColor={theme.colors.primaryAccent}
-          $contentType='our-background'
+          bgColor='var(--color-primary-accent)'
+          customInnerClass='our-background'
         />
       </RevealContentContainer>
 
@@ -72,9 +68,7 @@ export default function AboutUs() {
           rowOrder='row-reverse'
           image='/images/svg/open-angle-bracket.svg'
           altTag=''
-          color={theme.colors.primaryContent}
-          bgColor={theme.colors.white}
-          $contentType='our-process'
+          customInnerClass='peer-reviews'
         />
         <TwoColumn
           title='Version control'
@@ -94,9 +88,7 @@ export default function AboutUs() {
           }
           image='/images/svg/close-angle-bracket.svg'
           altTag=''
-          color={theme.colors.primaryContent}
-          bgColor={theme.colors.white}
-          $contentType='our-process'
+          customInnerClass='version-control'
         />
       </RevealContentContainer>
       <RevealContentContainer>
@@ -137,9 +129,7 @@ export default function AboutUs() {
           rowOrder='row-reverse'
           image='/images/svg/semi-colon.svg'
           altTag=''
-          color={theme.colors.primaryContent}
-          bgColor={theme.colors.white}
-          $contentType='our-purpose'
+          customInnerClass='our-purpose'
         />
       </RevealContentContainer>
       <RevealContentContainer>
@@ -148,9 +138,8 @@ export default function AboutUs() {
           content={''}
           image='/images/svg/close-curly-bracket.svg'
           altTag=''
-          color={theme.colors.primaryContent}
-          bgColor={theme.colors.lightBg}
-          $contentType='wanna-learn-more'
+          bgColor='var(--color-light-bg)'
+          customInnerClass='wanna-learn-more'
         />
         <TwoColumn
           title='Junior Developers'
@@ -165,9 +154,8 @@ export default function AboutUs() {
           openNewTab
           link='https://github.com/Web-Dev-Path/web-dev-path/wiki'
           customBtnClass='inverted-grey'
-          color={theme.colors.primaryContent}
-          bgColor={theme.colors.lightBg}
-          $contentType='two-text-columns'
+          bgColor='var(--color-light-bg)'
+          customInnerClass='two-text-columns'
           secondTextColumn={
             <TwoColumn
               title='Experienced Developers'
@@ -180,9 +168,8 @@ export default function AboutUs() {
               linkText='Contact us'
               link='/contact'
               customBtnClass='inverted-grey'
-              color={theme.colors.primaryContent}
-              bgColor={theme.colors.lightBg}
-              $contentType='second-text-column'
+              bgColor='var(--color-light-bg)'
+              customInnerClass='second-text-column'
             />
           }
         />
@@ -217,9 +204,7 @@ export default function AboutUs() {
           rowOrder='row-reverse'
           image='/images/svg/slash.svg'
           altTag=''
-          color={theme.colors.primaryContent}
-          bgColor={theme.colors.white}
-          $contentType='get-started'
+          customInnerClass='get-started'
           link='mailto:hello@webdevpath.co'
           linkText='Ping us'
           customBtnClass='inverted-grey'
@@ -254,12 +239,10 @@ export default function AboutUs() {
           content='Feel free to contact us.'
           image='/images/svg/hashtag.svg'
           altTag=''
-          color={theme.colors.primaryContent}
-          bgColor={theme.colors.white}
-          $contentType='questions'
           link='/contact'
           linkText='Contact us'
           customBtnClass='inverted-grey'
+          customInnerClass='questions'
         />
       </RevealContentContainer>
     </div>