+++ pages/index.js
@@ -1,11 +1,12 @@
 import TwoColumn from '@/components/containers/TwoColumn';
 import { CardsColumns } from '@/components/containers/CardColumns';
-import S from '@/styles/pages/homeStyles';
+import Bracket from '@/components/decorations/Bracket';
+import bracketStyles from '@/components/decorations/Bracket/Bracket.module.scss';
+import Stick from '@/components/decorations/Stick';
+import stickStyles from '@/components/decorations/Stick/Stick.module.scss';
 import RevealContentContainer from '@/components/containers/RevealContentContainer';
-import { useTheme } from 'styled-components';
 
 export default function Home() {
-  const theme = useTheme();
   return (
     <>
       <RevealContentContainer>
@@ -16,18 +17,18 @@ export default function Home() {
           content='Web Dev Path is an open-source initiative that provides hands-on experience in a simulated professional environment to people who seek to begin or move forward in their web development journey.'
           link='/about'
           customBtnClass='inverted-grey'
-          bgColor={theme.colors.lightBg}
+          bgColor='var(--color-light-bg)'
         />
       </RevealContentContainer>
 
       <RevealContentContainer>
-        <S.Bracket />
+        <Bracket className={bracketStyles.homeBracket} />
         <TwoColumn
           title='Get involved.'
           content='Web Dev Path runs on volunteers. Here are the ways you can get involved with us:'
           rowOrder='row-reverse'
-          $contentType='get-involved'
-          bgColor={theme.colors.white}
+          customInnerClass='get-involved'
+          bgColor='var(--color-white)'
         />
       </RevealContentContainer>
 
@@ -66,15 +67,14 @@ export default function Home() {
       </RevealContentContainer>
 
       <RevealContentContainer>
-        <S.Stick />
-
+        <Stick className={stickStyles.homeStick} />
         <TwoColumn
           title='Nonprofit?'
           content='Web Dev Path can help your nonprofit with web projects of various sizes. Connect with us to find out how.'
-          color={theme.colors.white}
-          bgColor={theme.colors.primaryContent}
+          color='var(--color-white)'
+          bgColor='var(--color-primary-content)'
           link='/about'
-          $contentType='non-profit'
+          customInnerClass='non-profit'
           customBtnClass='inverted-white'
         />
       </RevealContentContainer>